docker-run-app
I wrote a simple program to be the entry point of a docker container. It executes what would otherwise have been your container application.

More importantly, it configures handlers for POSIX signals and forwards them to your container application. I'm updating this description long after the asciicast was created, but I recall trying to use a bash shell script as the entry point of my container, but /bin/bash failed to respond to POSIX signals properly. So I wrote docker-run-app to do that job right.