How to compile from source
You could also run as root to avoid using sudo and getting the errors I did. ;-)

Note that not all source code has a 'make uninstall', this is why using something like a SlackBuild is better as it allows you to uninstall and track updates better.

Although ./configure was not necessary for 'st', it is necessary for most source tarballs. Usually your command set is:

$ ./configure $ make $ sudo make install

Some circles prefer:

$ make install clean

to 'make install', this will clear out the build directory after installing. Although not necessary, it is probably 'best' practice to use 'make install clean'.

P.S. The shortcut '!!' means to run the last command, prepending that with sudo saved me some typing. ;-)